// packages/shared/src/types/game/item-types.ts
export enum EquipmentSlotName {
  WEAPON = "weapon",
  SHIELD = "shield",
  HELMET = "helmet",
  BODY = "body",
  LEGS = "legs",
  BOOTS = "boots",
  GLOVES = "gloves",
  CAPE = "cape",
  AMULET = "amulet",
  RING = "ring",
  ARROWS = "arrows",
}